+++ slack-aws/commands.rb
@@ -1,2 +1,5 @@
 require 'slack-aws/commands/about'
 require 'slack-aws/commands/help'
+
+require 'aws-sdk'
+require 'slack-aws/commands/s3'