+++ spec/slack-aws/app_spec.rb
@@ -4,5 +4,7 @@ describe SlackAws::App do
   def app
     SlackAws::App.new
   end
-  it_behaves_like 'a slack ruby bot'
+  context 'authed' do
+    it_behaves_like 'a slack ruby bot'
+  end
 end