+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ### 0.1.0 (Next)
 
+* Added `aws s3 ls [bucket name]` - [@dblock](https://github.com/dblock).
+* Added `aws s3 list buckets` - [@dblock](https://github.com/dblock).
 * Initial public release - [@dblock](https://github.com/dblock).